+++ src/components/Header/Header.styled.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { AppBar, Button } from "@mui/material";
+import { AppBar } from "@mui/material";
 import { Link } from "react-router-dom";
 import { breakpoints } from "../../lib/breakpoints.ts";
 
@@ -22,19 +22,5 @@ export const Logo = styled(Link)`
 
   @media screen and (max-width: ${breakpoints.md}px) {
     font-size: 20px;
-  }
-`;
-
-export const ButtonStyled = styled(Button)`
-  background-color: #222020;
-  border-radius: 10px;
-  padding: 5px 50px;
-
-  &:hover {
-    background-color: #302f2f;
-  }
-
-  @media screen and (max-width: ${breakpoints.md}px) {
-    padding: 5px 25px;
   }
 `;
