+++ src/components/Header/Header.tsx
@@ -1,13 +1,12 @@
 import * as S from "./Header.styled";
+import Button from "../Button/Button.tsx";
 
 export default function Header() {
   return (
     <S.Header position="sticky">
       <S.Logo to="/">Some Company</S.Logo>
 
-      <S.ButtonStyled variant="contained" size="large">
-        Contact us
-      </S.ButtonStyled>
+      <Button title="Contact us" />
     </S.Header>
   );
 }
